tests/integration/gcp/requirements.txt
google-cloud-storage>=1.13.2